+++ src/index.js
@@ -3,21 +3,23 @@ import './sass/main.scss';
 import './js/components/pagination.js';
 import './js/apiService.js';
 
-import './js/markup-home.js';
+import './js/markupHome.js';
 import './js/markup-my-library.js';
 
 import './js/refs/refs.js';
 import './js/main.js';
 
-import './js/modal-team.js';
-import './js/components/button-up.js';
+import './js/modalTeam.js';
+import './js/components/buttonUp.js';
 
 import './js/layout/slider.js';
 import './sass/components/basicLightbox.min.css';
 import './js/components/basicLightbox.min.js';
 import './js/trailer.js';
 
-import './js/search-movie';
+import './js/searchMovie';
+
+import './js/components/buttonUp.js';
 
 import './js/theme-switch';
