+++ src/js/modal-film.js
@@ -94,9 +94,7 @@ function markupFilm(titleFilm) {
 function chekLocalStorage(dataFilm, keyLocal, nameBtn, refEl) {
   if (
     localStorage.getItem(keyLocal) &&
-    JSON.parse(localStorage.getItem(keyLocal)).some(
-      el => el.original_title === dataFilm.original_title || el.name === dataFilm.original_name,
-    )
+    JSON.parse(localStorage.getItem(keyLocal)).some(el => el.title === dataFilm.title)
   ) {
     refEl.textContent = `REMOVE FROM ${nameBtn}`;
   } else {
